requirements.txt
flask>=2.0.0
flask-socketio>=5.1.1
flask-cors>=3.0.10
eventlet>=0.33.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pyttsx3==2.98
SpeechRecognition==3.10.0
opencv-python==4.8.1.78
Pillow==10.0.0
numpy==1.24.3
google-cloud-texttospeech==2.14.1
psutil>=5.9.0
